requirements.txt

# FastAPI and web server
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM and embeddings
groq>=0.4.0
langchain>=0.0.267
langchain-groq>=0.0.5
langchain-community>=0.0.10

# Vector database
chromadb>=0.4.13

# Database
sqlalchemy>=2.0.0
alembic>=1.11.1
# sqlite3>=2.6.0

# Utilities
numpy>=1.24.3
pandas>=2.0.3
python-multipart>=0.0.6
tqdm>=4.65.0

# Testing
pytest>=7.3.1
httpx>=0.24.1

# Evaluation
ragas>=0.0.18
evaluate>=0.4.0